@@ -4,7 +4,6 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^4.6.2",
-    "@craco/craco": "^6.1.2",
     "@formspree/react": "^2.2.3",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
     "@headlessui/react": "^1.3.0",
@@ -12,6 +11,7 @@
     "@material-ui/core": "^4.11.4",
     "@material-ui/data-grid": "*",
     "@material-ui/icons": "^4.11.2",
+    "@project-serum/awesome-serum": "1.0.1",
     "@project-serum/borsh": "^0.2.1",
     "@project-serum/serum": "^0.13.34",
     "@pythnetwork/client": "^2.0.0",
@@ -34,11 +34,12 @@
     "dotenv": "^9.0.2",
     "embla-carousel": "^4.5.3",
     "humanize-duration-ts": "^2.1.1",
+    "immutable-tuple": "^0.4.10",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
     "notistack": "^1.0.9",
     "numeral": "2.0.6",
-    "postcss-import": "^14.0.1",
+    "postcss-import": "^7.0.1",
     "postcss-preset-env": "^7.2.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -46,24 +47,23 @@
     "react-icons": "^4.2.0",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^5.0.0",
+    "react-scripts": "^4.0.1",
     "react-simple-marquee": "^1.2.2",
     "react-toastify": "^7.0.4",
     "sass-loader": "^10",
+    "styled-components": "^5.1.1",
     "superstruct": "0.8.3",
     "types": "^0.1.1",
     "typescript": "^4.3.4",
-    "web-vitals": "^1.1.2",
-    "immutable-tuple": "^0.4.10",
-    "styled-components": "^5.1.1",
-    "@project-serum/awesome-serum": "1.0.1"
-
+    "typescript-map": "^0.1.0",
+    "web-vitals": "^1.1.2"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
-    "eject": "react-scripts eject"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "format": "prettier --write \"./src/**/*.{js,jsx,ts,tsx,json,css,scss,md}\""
   },
   "eslintConfig": {
     "extends": [
@@ -84,20 +84,32 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-transform-runtime": "^7.16.8",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
+    "@svgr/webpack": "^6.2.0",
     "@tailwindcss/postcss7-compat": "^2.2.2",
+    "@types/node": "17.0.8",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-virtualized": "^9.21.11",
+    "@types/styled-components": "^5.1.20",
     "@typescript-eslint/eslint-plugin": "^4.28.3",
     "@typescript-eslint/parser": "^4.28.3",
-    "autoprefixer": "^10.4.2",
+    "autoprefixer": "^9.4.2",
+    "babel-loader": "^8.2.3",
+    "clean-webpack-plugin": "4.0.0",
+    "copy-webpack-plugin": "^6.0.0",
     "eslint": "^7.31.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-unused-imports": "^1.1.2",
-    "postcss": "10.4.2",
+    "postcss": "^7.0.39",
+    "prettier": "^2.5.1",
     "react-async-hook": "^3.6.2",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.2",
-    "webpack": "^4.44.2",
-    "@types/styled-components": "^5.1.20"
+    "source-map-loader": "3.0.1",
+    "tailwindcss": "^3.0.14",
+    "webpack": "^5.46.0",
+    "webpack-bundle-analyzer": "4.5.0",
+    "webpack-cli": "^4.9.1"
   }
 }
