@@ -1,9 +1,5 @@
-export const Background = ({children}) => {
-    return(
-        <div className="bg__yyprime">
-            {children}
-        </div>
-    )
-};
+export const Background = ({ children }) => {
+  return <div className="bg__yyprime">{children}</div>
+}
 
-export default Background;
+export default Background
