@@ -1,27 +1,27 @@
-import React from "react";
-import { Typography } from "antd";
-import { shortenAddress } from "../../utils/utils";
-import { PublicKey } from "@solana/web3.js";
+import React from 'react'
+import { Typography } from 'antd'
+import { shortenAddress } from '../../utils/utils'
+import { PublicKey } from '@solana/web3.js'
 
 export const ExplorerLink = (props: {
-  address: string | PublicKey;
-  type: string;
-  code?: boolean;
-  style?: React.CSSProperties;
-  length?: number;
+  address: string | PublicKey
+  type: string
+  code?: boolean
+  style?: React.CSSProperties
+  length?: number
 }) => {
-  const { type, code } = props;
+  const { type, code } = props
 
   const address =
-    typeof props.address === "string"
+    typeof props.address === 'string'
       ? props.address
-      : props.address?.toBase58();
+      : props.address?.toBase58()
 
   if (!address) {
-    return null;
+    return null
   }
 
-  const length = props.length ?? 9;
+  const length = props.length ?? 9
 
   return (
     <a
@@ -29,7 +29,8 @@ export const ExplorerLink = (props: {
       // eslint-disable-next-line react/jsx-no-target-blank
       target="_blank"
       title={address}
-      style={props.style} rel="noreferrer"
+      style={props.style}
+      rel="noreferrer"
     >
       {code ? (
         <Typography.Text style={props.style} code>
@@ -39,5 +40,5 @@ export const ExplorerLink = (props: {
         shortenAddress(address, length)
       )}
     </a>
-  );
-};
+  )
+}
