@@ -6,14 +6,11 @@ const { Footer } = Layout;
 const { useBreakpoint } = Grid;
 
 const footerElements = [
-  {
-    description: 'Serum Developer Resources',
-    link: helpUrls.developerResources,
-  },
   { description: 'Discord', link: helpUrls.discord },
   { description: 'Telegram', link: helpUrls.telegram },
+  { description: 'Twitter', link: helpUrls.twitter },
   { description: 'GitHub', link: helpUrls.github },
-  { description: 'Project Serum', link: helpUrls.projectSerum },
+  { description: 'Yyprime Exchange', link: helpUrls.yyprime },
   { description: 'Solana Network', link: helpUrls.solanaBeach },
 ];
 
