@@ -1,25 +1,25 @@
-import React from 'react';
-import { PublicKey } from '@solana/web3.js';
-import { abbreviateAddress } from '../utils/utilsSerum';
-import { useMintToTickers } from '../utils/tokens';
-import { Popover } from 'antd';
-import LinkAddress from './LinkAddress';
-import { InfoCircleOutlined } from '@ant-design/icons';
+import React from 'react'
+import { PublicKey } from '@solana/web3.js'
+import { abbreviateAddress } from '../utils/utilsSerum'
+import { useMintToTickers } from '../utils/tokens'
+import { Popover } from 'antd'
+import LinkAddress from './LinkAddress'
+import { InfoCircleOutlined } from '@ant-design/icons'
 
 export function MintName({
   mint,
   showAddress = false,
 }: {
-  mint: string | PublicKey | null | undefined;
-  showAddress?: boolean;
+  mint: string | PublicKey | null | undefined
+  showAddress?: boolean
 }) {
-  const mintToTickers = useMintToTickers();
+  const mintToTickers = useMintToTickers()
   if (!mint) {
-    return null;
+    return null
   }
-  const mintKey = typeof mint === 'string' ? new PublicKey(mint) : mint;
-  const mintAddress = typeof mint === 'string' ? mint : mint.toBase58();
-  const ticker = mintToTickers[mintAddress] ?? abbreviateAddress(mintKey);
+  const mintKey = typeof mint === 'string' ? new PublicKey(mint) : mint
+  const mintAddress = typeof mint === 'string' ? mint : mint.toBase58()
+  const ticker = mintToTickers[mintAddress] ?? abbreviateAddress(mintKey)
 
   return (
     <>
@@ -38,5 +38,5 @@ export function MintName({
         </>
       ) : null}
     </>
-  );
+  )
 }
