@@ -1,11 +1,11 @@
-import React from 'react';
-import styled from 'styled-components';
+import React from 'react'
+import styled from 'styled-components'
 
 const Wrapper = styled.div`
   margin: 5px;
   padding: 20px;
   background-color: #1a2029;
-`;
+`
 
 export default function FloatingElement({
   style = undefined,
@@ -21,5 +21,5 @@ export default function FloatingElement({
     >
       {children}
     </Wrapper>
-  );
+  )
 }
