@@ -1,4 +1,4 @@
-import { useLocalStorageState } from "./../utils/utils";
+import { useLocalStorageState } from './../utils/utils'
 import {
   Account,
   clusterApiUrl,
@@ -6,57 +6,57 @@ import {
   PublicKey,
   Transaction,
   TransactionInstruction,
-} from "@solana/web3.js";
-import React, { useContext, useEffect, useMemo, useState } from "react";
-import { notify } from "./../utils/notifications";
-import { ExplorerLink } from "../components/ExplorerLink";
-import { setProgramIds } from "../utils/ids";
+} from '@solana/web3.js'
+import React, { useContext, useEffect, useMemo, useState } from 'react'
+import { notify } from './../utils/notifications'
+import { ExplorerLink } from '../components/ExplorerLink'
+import { setProgramIds } from '../utils/ids'
 // import { WalletAdapter } from "@saberhq/use-solana";
-import { cache, getMultipleAccounts, MintParser } from "./accounts";
-import { TokenListProvider, ENV as ChainID, TokenInfo } from "@solana/spl-token-registry";
+import { cache, getMultipleAccounts, MintParser } from './accounts'
+import {
+  TokenListProvider,
+  ENV as ChainID,
+  TokenInfo,
+} from '@solana/spl-token-registry'
 
-export type ENV =
-  | "mainnet-beta"
-  | "testnet"
-  | "devnet"
-  | "localnet";
+export type ENV = 'mainnet-beta' | 'testnet' | 'devnet' | 'localnet'
 
 export const ENDPOINTS = [
   {
-    name: "mainnet-beta" as ENV,
-    endpoint: "https://solana-api.projectserum.com/",
+    name: 'mainnet-beta' as ENV,
+    endpoint: 'https://solana-api.projectserum.com/',
     chainID: ChainID.MainnetBeta,
   },
   {
-    name: "testnet" as ENV,
-    endpoint: clusterApiUrl("testnet"),
+    name: 'testnet' as ENV,
+    endpoint: clusterApiUrl('testnet'),
     chainID: ChainID.Testnet,
   },
   {
-    name: "devnet" as ENV,
-    endpoint: clusterApiUrl("devnet"),
+    name: 'devnet' as ENV,
+    endpoint: clusterApiUrl('devnet'),
     chainID: ChainID.Devnet,
   },
   {
-    name: "localnet" as ENV,
-    endpoint: "http://127.0.0.1:8899",
+    name: 'localnet' as ENV,
+    endpoint: 'http://127.0.0.1:8899',
     chainID: ChainID.Devnet,
   },
-];
+]
 
-const DEFAULT = ENDPOINTS[2].endpoint;
-const DEFAULT_SLIPPAGE = 0.25;
+const DEFAULT = ENDPOINTS[2].endpoint
+const DEFAULT_SLIPPAGE = 0.25
 
 interface ConnectionConfig {
-  connection: Connection;
-  sendConnection: Connection;
-  endpoint: string;
-  slippage: number;
-  setSlippage: (val: number) => void;
-  env: ENV;
-  setEndpoint: (val: string) => void;
-  tokens: TokenInfo[];
-  tokenMap: Map<string, TokenInfo>;
+  connection: Connection
+  sendConnection: Connection
+  endpoint: string
+  slippage: number
+  setSlippage: (val: number) => void
+  env: ENV
+  setEndpoint: (val: string) => void
+  tokens: TokenInfo[]
+  tokenMap: Map<string, TokenInfo>
 }
 
 const ConnectionContext = React.createContext<ConnectionConfig>({
@@ -64,101 +64,104 @@ const ConnectionContext = React.createContext<ConnectionConfig>({
   setEndpoint: () => {},
   slippage: DEFAULT_SLIPPAGE,
   setSlippage: (val: number) => {},
-  connection: new Connection(DEFAULT, "recent"),
-  sendConnection: new Connection(DEFAULT, "recent"),
+  connection: new Connection(DEFAULT, 'recent'),
+  sendConnection: new Connection(DEFAULT, 'recent'),
   env: ENDPOINTS[2].name,
   tokens: [],
   tokenMap: new Map<string, TokenInfo>(),
-});
+})
 
 export function ConnectionProvider({ children = undefined as any }) {
   const [endpoint, setEndpoint] = useLocalStorageState(
-    "connectionEndpts",
+    'connectionEndpts',
     ENDPOINTS[2].endpoint
-  );
+  )
 
   const [slippage, setSlippage] = useLocalStorageState(
-    "slippage",
+    'slippage',
     DEFAULT_SLIPPAGE.toString()
-  );
+  )
 
-  const connection = useMemo(() => new Connection(endpoint, "recent"), [
-    endpoint,
-  ]);
-  const sendConnection = useMemo(() => new Connection(endpoint, "recent"), [
-    endpoint,
-  ]);
+  const connection = useMemo(
+    () => new Connection(endpoint, 'recent'),
+    [endpoint]
+  )
+  const sendConnection = useMemo(
+    () => new Connection(endpoint, 'recent'),
+    [endpoint]
+  )
 
   const chain =
-    ENDPOINTS.find((end) => end.endpoint === endpoint) || ENDPOINTS[2];
-  const env = chain.name;
+    ENDPOINTS.find((end) => end.endpoint === endpoint) || ENDPOINTS[2]
+  const env = chain.name
 
-  const [tokens, setTokens] = useState<TokenInfo[]>([]);
-  const [tokenMap, setTokenMap] = useState<Map<string, TokenInfo>>(new Map());
+  const [tokens, setTokens] = useState<TokenInfo[]>([])
+  const [tokenMap, setTokenMap] = useState<Map<string, TokenInfo>>(new Map())
   useEffect(() => {
-    cache.clear();
+    cache.clear()
     // fetch token files
-    (async () => {
-      const res = await new TokenListProvider().resolve();
+    ;(async () => {
+      const res = await new TokenListProvider().resolve()
       const list = res
         .filterByChainId(chain.chainID)
-        .excludeByTag("nft")
-        .getList();
+        .excludeByTag('nft')
+        .getList()
       const knownMints = list.reduce((map, item) => {
-        map.set(item.address, item);
-        return map;
-      }, new Map<string, TokenInfo>());
+        map.set(item.address, item)
+        return map
+      }, new Map<string, TokenInfo>())
 
-      const accounts = await getMultipleAccounts(connection, [...knownMints.keys()], 'single');
+      const accounts = await getMultipleAccounts(
+        connection,
+        [...knownMints.keys()],
+        'single'
+      )
       accounts.keys.forEach((key, index) => {
-        const account = accounts.array[index];
-        if(!account) {
-          return;
+        const account = accounts.array[index]
+        if (!account) {
+          return
         }
 
-        cache.add(new PublicKey(key), account, MintParser);
+        cache.add(new PublicKey(key), account, MintParser)
       })
 
-      setTokenMap(knownMints);
-      setTokens(list);
-    })();
-  }, [connection, chain]);
+      setTokenMap(knownMints)
+      setTokens(list)
+    })()
+  }, [connection, chain])
 
-  setProgramIds(env);
+  setProgramIds(env)
 
   // The websocket library solana/web3.js uses closes its websocket connection when the subscription list
   // is empty after opening its first time, preventing subsequent subscriptions from receiving responses.
   // This is a hack to prevent the list from every getting empty
   useEffect(() => {
-    const id = connection.onAccountChange(new Account().publicKey, () => {});
+    const id = connection.onAccountChange(new Account().publicKey, () => {})
     return () => {
-      connection.removeAccountChangeListener(id);
-    };
-  }, [connection]);
+      connection.removeAccountChangeListener(id)
+    }
+  }, [connection])
 
   useEffect(() => {
-    const id = connection.onSlotChange(() => null);
+    const id = connection.onSlotChange(() => null)
     return () => {
-      connection.removeSlotChangeListener(id);
-    };
-  }, [connection]);
+      connection.removeSlotChangeListener(id)
+    }
+  }, [connection])
 
   useEffect(() => {
-    const id = sendConnection.onAccountChange(
-      new Account().publicKey,
-      () => {}
-    );
+    const id = sendConnection.onAccountChange(new Account().publicKey, () => {})
     return () => {
-      sendConnection.removeAccountChangeListener(id);
-    };
-  }, [sendConnection]);
+      sendConnection.removeAccountChangeListener(id)
+    }
+  }, [sendConnection])
 
   useEffect(() => {
-    const id = sendConnection.onSlotChange(() => null);
+    const id = sendConnection.onSlotChange(() => null)
     return () => {
-      sendConnection.removeSlotChangeListener(id);
-    };
-  }, [sendConnection]);
+      sendConnection.removeSlotChangeListener(id)
+    }
+  }, [sendConnection])
 
   return (
     <ConnectionContext.Provider
@@ -176,72 +179,72 @@ export function ConnectionProvider({ children = undefined as any }) {
     >
       {children}
     </ConnectionContext.Provider>
-  );
+  )
 }
 
 export function useConnection() {
-  return useContext(ConnectionContext).connection as Connection;
+  return useContext(ConnectionContext).connection as Connection
 }
 
 export function useSendConnection() {
-  return useContext(ConnectionContext)?.sendConnection;
+  return useContext(ConnectionContext)?.sendConnection
 }
 
 export function useConnectionConfig() {
-  const context = useContext(ConnectionContext);
+  const context = useContext(ConnectionContext)
   return {
     endpoint: context.endpoint,
     setEndpoint: context.setEndpoint,
     env: context.env,
     tokens: context.tokens,
     tokenMap: context.tokenMap,
-  };
+  }
 }
 
 export function useSlippageConfig() {
-  const { slippage, setSlippage } = useContext(ConnectionContext);
-  return { slippage, setSlippage };
+  const { slippage, setSlippage } = useContext(ConnectionContext)
+  return { slippage, setSlippage }
 }
 
 const getErrorForTransaction = async (connection: Connection, txid: string) => {
   // wait for all confirmation before geting transaction
-  await connection.confirmTransaction(txid, "max");
+  await connection.confirmTransaction(txid, 'max')
 
-  const tx = await connection.getParsedConfirmedTransaction(txid);
+  const tx = await connection.getParsedConfirmedTransaction(txid)
 
-  const errors: string[] = [];
+  const errors: string[] = []
   if (tx?.meta && tx.meta.logMessages) {
     tx.meta.logMessages.forEach((log) => {
-      const regex = /Error: (.*)/gm;
-      let m;
+      const regex = /Error: (.*)/gm
+      let m
       while ((m = regex.exec(log)) !== null) {
         // This is necessary to avoid infinite loops with zero-width matches
         if (m.index === regex.lastIndex) {
-          regex.lastIndex++;
+          regex.lastIndex++
         }
 
         if (m.length > 1) {
-          errors.push(m[1]);
+          errors.push(m[1])
         }
       }
-    });
+    })
   }
 
-  return errors;
-};
+  return errors
+}
 
 export function useSolanaExplorerUrlSuffix() {
-  const context = useContext(ConnectionContext);
-    if (!context) {
-      throw new Error('Missing connection context');
-    }
-    const endpoint = context.endpoint;
-      if (endpoint === clusterApiUrl('devnet')) {
-        return '?cluster=devnet';
-      } else if (endpoint === clusterApiUrl('testnet')) {
-        return '?cluster=testnet';
-      }
-    return '';
+  const context = useContext(ConnectionContext)
+  if (!context) {
+    throw new Error('Missing connection context')
+  }
+  const endpoint = context.endpoint
+  if (endpoint === clusterApiUrl('devnet')) {
+    return '?cluster=devnet'
+  } else if (endpoint === clusterApiUrl('testnet')) {
+    return '?cluster=testnet'
+  }
+  return ''
 }
 
 // export const sendTransaction = async (
