@@ -1,9 +1,9 @@
-import { TokenAccount } from "../models";
-import { useAccountsContext } from "./../contexts/accounts";
+import { TokenAccount } from '../models'
+import { useAccountsContext } from './../contexts/accounts'
 
 export function useUserAccounts() {
-  const context = useAccountsContext();
+  const context = useAccountsContext()
   return {
     userAccounts: context.userAccounts as TokenAccount[],
-  };
+  }
 }
