@@ -1,24 +1,22 @@
-import { HashRouter, Route, Switch, Redirect } from 'react-router-dom';
-import TradePage from './views/TradePage';
+import { HashRouter, Route, Switch, Redirect } from 'react-router-dom'
+import TradePage from './views/TradePage'
 
-import React from 'react';
+import React from 'react'
 
-import BasicLayout from './components/BasicLayout';
+import BasicLayout from './components/BasicLayout'
 
-import { getTradePageUrl } from './utils/markets';
-import Simulator from './views';
-
+import { getTradePageUrl } from './utils/markets'
+import Simulator from './views'
 
 export function Routes() {
   return (
     <>
-    <Route exact path="/simulator">
-              <Simulator />
-            </Route>
       <HashRouter basename={'/'}>
         <BasicLayout>
           <Switch>
-          
+            <Route exact path="/simulator">
+              <Simulator />
+            </Route>
             <Route exact path="/market/:marketAddress">
               <TradePage />
             </Route>
@@ -26,5 +24,5 @@ export function Routes() {
         </BasicLayout>
       </HashRouter>
     </>
-  );
+  )
 }
