@@ -1,9 +1,9 @@
-declare module "buffer-layout" {
-  const bl: any;
-  export = bl;
+declare module 'buffer-layout' {
+  const bl: any
+  export = bl
 }
 
-declare module "jazzicon" {
-  const jazzicon: any;
-  export = jazzicon;
+declare module 'jazzicon' {
+  const jazzicon: any
+  export = jazzicon
 }
