@@ -1,26 +1,26 @@
-import { BonfidaTrade } from './types';
+import { BonfidaTrade } from './types'
 
 export default class BonfidaApi {
-  static URL: string = 'https://serum-api.bonfida.com/';
+  static URL: string = 'https://serum-api.bonfida.com/'
 
   static async get(path: string) {
     try {
-      const response = await fetch(this.URL + path);
+      const response = await fetch(this.URL + path)
       if (response.ok) {
-        const responseJson = await response.json();
-        return responseJson.success ? responseJson.data : null;
+        const responseJson = await response.json()
+        return responseJson.success ? responseJson.data : null
       }
     } catch (err) {
-      console.log(`Error fetching from Bonfida API ${path}: ${err}`);
+      console.log(`Error fetching from Bonfida API ${path}: ${err}`)
     }
-    return null;
+    return null
   }
 
   static async getRecentTrades(
-    marketAddress: string,
+    marketAddress: string
   ): Promise<BonfidaTrade[] | null> {
-    return BonfidaApi.get(`trades/address/${marketAddress}`);
+    return BonfidaApi.get(`trades/address/${marketAddress}`)
   }
 }
 
-export const BONFIDA_DATA_FEED = 'https://serum-api.bonfida.com/tv';
+export const BONFIDA_DATA_FEED = 'https://serum-api.bonfida.com/tv'
