@@ -1,18 +1,18 @@
-import { useEffect, useReducer } from 'react';
+import { useEffect, useReducer } from 'react'
 
-import assert from 'assert';
+import assert from 'assert'
 
-const pageLoadTime = new Date();
+const pageLoadTime = new Date()
 
-const globalCache: Map<any, any> = new Map();
+const globalCache: Map<any, any> = new Map()
 
 class FetchLoopListener<T = any> {
-  cacheKey: any;
-  fn: () => Promise<T>;
-  refreshInterval: number;
-  refreshIntervalOnError: number | null;
-  callback: () => void;
-  cacheNullValues: Boolean = true;
+  cacheKey: any
+  fn: () => Promise<T>
+  refreshInterval: number
+  refreshIntervalOnError: number | null
+  callback: () => void
+  cacheNullValues: Boolean = true
 
   constructor(
     cacheKey: any,
@@ -20,143 +20,143 @@ class FetchLoopListener<T = any> {
     refreshInterval: number,
     refreshIntervalOnError: number | null,
     callback: () => void,
-    cacheNullValues: Boolean,
+    cacheNullValues: Boolean
   ) {
-    this.cacheKey = cacheKey;
-    this.fn = fn;
-    this.refreshInterval = refreshInterval;
-    this.refreshIntervalOnError = refreshIntervalOnError;
-    this.callback = callback;
-    this.cacheNullValues = cacheNullValues;
+    this.cacheKey = cacheKey
+    this.fn = fn
+    this.refreshInterval = refreshInterval
+    this.refreshIntervalOnError = refreshIntervalOnError
+    this.callback = callback
+    this.cacheNullValues = cacheNullValues
   }
 }
 
 class FetchLoopInternal<T = any> {
-  cacheKey: any;
-  fn: () => Promise<T>;
-  timeoutId: null | any;
-  listeners: Set<FetchLoopListener<T>>;
-  errors: number;
-  cacheNullValues: Boolean = true;
+  cacheKey: any
+  fn: () => Promise<T>
+  timeoutId: null | any
+  listeners: Set<FetchLoopListener<T>>
+  errors: number
+  cacheNullValues: Boolean = true
 
   constructor(cacheKey: any, fn: () => Promise<T>, cacheNullValues: Boolean) {
-    this.cacheKey = cacheKey;
-    this.fn = fn;
-    this.timeoutId = null;
-    this.listeners = new Set();
-    this.errors = 0;
-    this.cacheNullValues = cacheNullValues;
+    this.cacheKey = cacheKey
+    this.fn = fn
+    this.timeoutId = null
+    this.listeners = new Set()
+    this.errors = 0
+    this.cacheNullValues = cacheNullValues
   }
 
   get refreshInterval(): number {
     return Math.min(
-      ...[...this.listeners].map((listener) => listener.refreshInterval),
-    );
+      ...[...this.listeners].map((listener) => listener.refreshInterval)
+    )
   }
 
   get refreshIntervalOnError(): number | null {
     const refreshIntervalsOnError: number[] = [...this.listeners]
       .map((listener) => listener.refreshIntervalOnError)
-      .filter((x): x is number => x !== null);
+      .filter((x): x is number => x !== null)
     if (refreshIntervalsOnError.length === 0) {
-      return null;
+      return null
     }
-    return Math.min(...refreshIntervalsOnError);
+    return Math.min(...refreshIntervalsOnError)
   }
 
   get stopped(): boolean {
-    return this.listeners.size === 0;
+    return this.listeners.size === 0
   }
 
   addListener(listener: FetchLoopListener<T>): void {
-    const previousRefreshInterval = this.refreshInterval;
-    this.listeners.add(listener);
+    const previousRefreshInterval = this.refreshInterval
+    this.listeners.add(listener)
     if (this.refreshInterval < previousRefreshInterval) {
-      this.refresh();
+      this.refresh()
     }
   }
 
   removeListener(listener: FetchLoopListener<T>): void {
-    assert(this.listeners.delete(listener));
+    assert(this.listeners.delete(listener))
     if (this.stopped) {
       if (this.timeoutId) {
-        clearTimeout(this.timeoutId);
-        this.timeoutId = null;
+        clearTimeout(this.timeoutId)
+        this.timeoutId = null
       }
     }
   }
 
   notifyListeners(): void {
-    this.listeners.forEach((listener) => listener.callback());
+    this.listeners.forEach((listener) => listener.callback())
   }
 
   refresh = async () => {
     if (this.timeoutId) {
-      clearTimeout(this.timeoutId);
-      this.timeoutId = null;
+      clearTimeout(this.timeoutId)
+      this.timeoutId = null
     }
     if (this.stopped) {
-      return;
+      return
     }
 
-    let errored = false;
+    let errored = false
     try {
-      const data = await this.fn();
+      const data = await this.fn()
       if (!this.cacheNullValues && data === null) {
-        console.log(`Not caching null value for ${this.cacheKey}`);
+        console.log(`Not caching null value for ${this.cacheKey}`)
         // cached data has not changed so no need to re-render
-        this.errors = 0;
-        return data;
+        this.errors = 0
+        return data
       } else {
-        globalCache.set(this.cacheKey, data);
-        this.errors = 0;
-        this.notifyListeners();
-        return data;
+        globalCache.set(this.cacheKey, data)
+        this.errors = 0
+        this.notifyListeners()
+        return data
       }
     } catch (error) {
-      ++this.errors;
-      console.warn(error);
-      errored = true;
+      ++this.errors
+      console.warn(error)
+      errored = true
     } finally {
       if (!this.timeoutId && !this.stopped) {
-        let waitTime = this.refreshInterval;
+        let waitTime = this.refreshInterval
         if (
           errored &&
           this.refreshIntervalOnError &&
           this.refreshIntervalOnError > 0
         ) {
-          waitTime = this.refreshIntervalOnError;
+          waitTime = this.refreshIntervalOnError
         }
 
         // Back off on errors.
         if (this.errors > 0) {
-          waitTime = Math.min(1000 * 2 ** (this.errors - 1), 60000);
+          waitTime = Math.min(1000 * 2 ** (this.errors - 1), 60000)
         }
 
         // Don't do any refreshing for the first five seconds, to make way for other things to load.
-        const timeSincePageLoad = +new Date() - +pageLoadTime;
+        const timeSincePageLoad = +new Date() - +pageLoadTime
         if (timeSincePageLoad < 5000) {
-          waitTime += 5000 - timeSincePageLoad / 2;
+          waitTime += 5000 - timeSincePageLoad / 2
         }
 
         // Refresh background pages slowly.
         if (document.visibilityState === 'hidden') {
-          waitTime = 60000;
+          waitTime = 60000
         } else if (!document.hasFocus()) {
-          waitTime *= 1.5;
+          waitTime *= 1.5
         }
 
         // Add jitter so we don't send all requests at the same time.
-        waitTime *= 0.8 + 0.4 * Math.random();
+        waitTime *= 0.8 + 0.4 * Math.random()
 
-        this.timeoutId = setTimeout(this.refresh, waitTime);
+        this.timeoutId = setTimeout(this.refresh, waitTime)
       }
     }
-  };
+  }
 }
 
 class FetchLoops {
-  loops = new Map();
+  loops = new Map()
 
   addListener<T>(listener: FetchLoopListener<T>) {
     if (!this.loops.has(listener.cacheKey)) {
@@ -165,46 +165,46 @@ class FetchLoops {
         new FetchLoopInternal<T>(
           listener.cacheKey,
           listener.fn,
-          listener.cacheNullValues,
-        ),
-      );
+          listener.cacheNullValues
+        )
+      )
     }
-    this.loops.get(listener.cacheKey).addListener(listener);
+    this.loops.get(listener.cacheKey).addListener(listener)
   }
 
   removeListener<T>(listener: FetchLoopListener<T>) {
-    const loop = this.loops.get(listener.cacheKey);
-    loop.removeListener(listener);
+    const loop = this.loops.get(listener.cacheKey)
+    loop.removeListener(listener)
     if (loop.stopped) {
-      this.loops.delete(listener.cacheKey);
-      globalCache.delete(listener.cacheKey);
+      this.loops.delete(listener.cacheKey)
+      globalCache.delete(listener.cacheKey)
     }
   }
 
   refresh(cacheKey) {
     if (this.loops.has(cacheKey)) {
-      this.loops.get(cacheKey).refresh();
+      this.loops.get(cacheKey).refresh()
     }
   }
 
   refreshAll() {
-    return Promise.all([...this.loops.values()].map((loop) => loop.refresh()));
+    return Promise.all([...this.loops.values()].map((loop) => loop.refresh()))
   }
 }
-const globalLoops = new FetchLoops();
+const globalLoops = new FetchLoops()
 
 export function useAsyncData<T = any>(
   asyncFn: () => Promise<T>,
   cacheKey: any,
   { refreshInterval = 60000, refreshIntervalOnError = null } = {},
-  cacheNullValues: Boolean = true,
+  cacheNullValues: Boolean = true
 ): [null | undefined | T, boolean] {
-  const [, rerender] = useReducer((i) => i + 1, 0);
+  const [, rerender] = useReducer((i) => i + 1, 0)
 
   useEffect(() => {
     if (!cacheKey) {
       // eslint-disable-next-line @typescript-eslint/no-empty-function
-      return () => {};
+      return () => {}
     }
     const listener = new FetchLoopListener<T>(
       cacheKey,
@@ -212,58 +212,56 @@ export function useAsyncData<T = any>(
       refreshInterval,
       refreshIntervalOnError,
       rerender,
-      cacheNullValues,
-    );
-    globalLoops.addListener(listener);
-    return () => globalLoops.removeListener(listener);
+      cacheNullValues
+    )
+    globalLoops.addListener(listener)
+    return () => globalLoops.removeListener(listener)
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [cacheKey, refreshInterval]);
+  }, [cacheKey, refreshInterval])
 
   if (!cacheKey) {
-    return [null, false];
+    return [null, false]
   }
 
-  const loaded = globalCache.has(cacheKey);
-  const data = loaded ? globalCache.get(cacheKey) : undefined;
-  return [data, loaded];
+  const loaded = globalCache.has(cacheKey)
+  const data = loaded ? globalCache.get(cacheKey) : undefined
+  return [data, loaded]
 }
 
 export function refreshCache(cacheKey: any, clearCache = false): void {
   if (clearCache) {
-    globalCache.delete(cacheKey);
+    globalCache.delete(cacheKey)
   }
-  const loop = globalLoops.loops.get(cacheKey);
+  const loop = globalLoops.loops.get(cacheKey)
   if (loop) {
-    loop.refresh();
+    loop.refresh()
     if (clearCache) {
-      loop.notifyListeners();
+      loop.notifyListeners()
     }
   }
 }
 
 export function refreshAllCaches(): void {
   for (const loop of globalLoops.loops.values()) {
-    loop.refresh();
+    loop.refresh()
   }
 }
 
 export function setCache(
   cacheKey: any,
   value: any,
-  { initializeOnly = false } = {},
+  { initializeOnly = false } = {}
 ): void {
   if (initializeOnly && globalCache.has(cacheKey)) {
-    return;
+    return
   }
-  globalCache.set(cacheKey, value);
-  const loop = globalLoops.loops.get(cacheKey);
+  globalCache.set(cacheKey, value)
+  const loop = globalLoops.loops.get(cacheKey)
   if (loop) {
-    loop.notifyListeners();
+    loop.notifyListeners()
   }
 }
 
-export function getCache(
-  cacheKey: any
-) {
-  return globalCache.get(cacheKey);
+export function getCache(cacheKey: any) {
+  return globalCache.get(cacheKey)
 }
