@@ -1,47 +1,48 @@
-import { PublicKey } from "@solana/web3.js";
-// For Identifying Stake Program Code 
-export const STAKE_PROGRAM_ID = new PublicKey('Stake11111111111111111111111111111111111111');
+import { PublicKey } from '@solana/web3.js'
+// For Identifying Stake Program Code
+export const STAKE_PROGRAM_ID = new PublicKey(
+  'Stake11111111111111111111111111111111111111'
+)
 
 export const WRAPPED_SOL_MINT = new PublicKey(
-  "So11111111111111111111111111111111111111112"
-);
+  'So11111111111111111111111111111111111111112'
+)
 export let TOKEN_PROGRAM_ID = new PublicKey(
-  "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
-);
+  'TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA'
+)
 
 export let LENDING_PROGRAM_ID = new PublicKey(
-  "TokenLending1111111111111111111111111111111"
-);
+  'TokenLending1111111111111111111111111111111'
+)
 
 export let SWAP_PROGRAM_ID = new PublicKey(
-  "SwaPpA9LAaLfeLi3a68M4DjnLqgtticKg6CnyNwgAC8"
-);
+  'SwaPpA9LAaLfeLi3a68M4DjnLqgtticKg6CnyNwgAC8'
+)
 
 export const PROGRAM_IDS = [
-    {
-      name: "mainnet-beta",
-    },
-    {
-      name: "testnet",
-    },
-    {
-      name: "devnet",
-    },
-    {
-      name: "localnet",
-    },
-  ];
-  
-  export const setProgramIds = (envName: string) => {
-    let instance = PROGRAM_IDS.find((env) => env.name === envName);
-    if (!instance) {
-      return;
-    }
-  };
+  {
+    name: 'mainnet-beta',
+  },
+  {
+    name: 'testnet',
+  },
+  {
+    name: 'devnet',
+  },
+  {
+    name: 'localnet',
+  },
+]
 
-  export const programIds = () => {
-    return {
-      token: TOKEN_PROGRAM_ID,
-    };
-  };
-  
+export const setProgramIds = (envName: string) => {
+  let instance = PROGRAM_IDS.find((env) => env.name === envName)
+  if (!instance) {
+    return
+  }
+}
+
+export const programIds = () => {
+  return {
+    token: TOKEN_PROGRAM_ID,
+  }
+}
