@@ -1,66 +1,72 @@
-import { Button } from '@material-ui/core';
-import { Connection } from '@solana/web3.js';
-import { useSnackbar } from 'notistack';
-import { useState } from 'react';
-import { notification } from "antd";
+import { Button } from '@material-ui/core'
+import { Connection } from '@solana/web3.js'
+import { useSnackbar } from 'notistack'
+import { useState } from 'react'
+import { notification } from 'antd'
 // import Link from '../components/Link';
-import { useConnection, useSolanaExplorerUrlSuffix } from '../contexts/connection';
+import {
+  useConnection,
+  useSolanaExplorerUrlSuffix,
+} from '../contexts/connection'
 
-async function confirmTransaction(
-  connection: Connection,
-  signature: string,
-) {
-  let startTime = new Date();
-  let result = await connection.confirmTransaction(signature, 'confirmed');
+async function confirmTransaction(connection: Connection, signature: string) {
+  let startTime = new Date()
+  let result = await connection.confirmTransaction(signature, 'confirmed')
   if (result.value.err) {
     throw new Error(
-      'Error confirming transaction: ' + JSON.stringify(result.value.err),
-    );
+      'Error confirming transaction: ' + JSON.stringify(result.value.err)
+    )
   }
   console.log(
     'Transaction confirmed after %sms',
-    new Date().getTime() - startTime.getTime(),
-  );
-  return result.value;
+    new Date().getTime() - startTime.getTime()
+  )
+  return result.value
 }
 
 export function useMonitorTransaction() {
-  const connection = useConnection();
-  const { enqueueSnackbar, closeSnackbar } = useSnackbar();
-  const [sending, setSending] = useState(false);
+  const connection = useConnection()
+  const { enqueueSnackbar, closeSnackbar } = useSnackbar()
+  const [sending, setSending] = useState(false)
 
   async function monitorTransaction(
     signaturePromise: Promise<string>,
-    {onSuccess, onError}: { onSuccess: (signature: string) => void, onError: (signature: string) => void },
+    {
+      onSuccess,
+      onError,
+    }: {
+      onSuccess: (signature: string) => void
+      onError: (signature: string) => void
+    }
   ): Promise<void> {
     let id = enqueueSnackbar('Sending transaction...', {
       variant: 'info',
       persist: true,
-    });
-    setSending(true);
+    })
+    setSending(true)
     try {
-      let signature = await signaturePromise;
-      closeSnackbar(id);
+      let signature = await signaturePromise
+      closeSnackbar(id)
       id = enqueueSnackbar('Confirming transaction...', {
         variant: 'info',
         persist: true,
         action: <ViewTransactionOnExplorerButton signature={signature} />,
-      });
-      await confirmTransaction(connection, signature);
-      closeSnackbar(id);
-      setSending(false);
+      })
+      await confirmTransaction(connection, signature)
+      closeSnackbar(id)
+      setSending(false)
       enqueueSnackbar('Transaction confirmed', {
         variant: 'success',
         autoHideDuration: 15000,
         action: <ViewTransactionOnExplorerButton signature={signature} />,
-      });
+      })
       if (onSuccess) {
-        onSuccess(signature);
+        onSuccess(signature)
       }
     } catch (e) {
-      closeSnackbar(id);
-      setSending(false);
-      console.warn(e);
+      closeSnackbar(id)
+      setSending(false)
+      console.warn(e)
       // enqueueSnackbar(e.message, { variant: 'error' });
       // if (onError) {
       //   onError(e);
@@ -68,11 +74,11 @@ export function useMonitorTransaction() {
     }
   }
 
-  return {monitorTransaction, sending};
+  return { monitorTransaction, sending }
 }
 
-function ViewTransactionOnExplorerButton({ signature }: {signature: string}) {
-  const urlSuffix = useSolanaExplorerUrlSuffix();
+function ViewTransactionOnExplorerButton({ signature }: { signature: string }) {
+  const urlSuffix = useSolanaExplorerUrlSuffix()
   return (
     <Button
       color="inherit"
@@ -83,11 +89,11 @@ function ViewTransactionOnExplorerButton({ signature }: {signature: string}) {
     >
       View on Explorer
     </Button>
-  );
+  )
 }
 
 export function useCallAsync() {
-  const { enqueueSnackbar, closeSnackbar } = useSnackbar();
+  const { enqueueSnackbar, closeSnackbar } = useSnackbar()
   return async function callAsync(
     promise: Promise<string>,
     {
@@ -96,44 +102,42 @@ export function useCallAsync() {
       onSuccess,
       onError,
     }: {
-      progressMessage: string,
-      successMessage: string,
-      onSuccess: (message: string) => void,
-      onError: (message: string) => void,
+      progressMessage: string
+      successMessage: string
+      onSuccess: (message: string) => void
+      onError: (message: string) => void
     }
   ) {
     let id = enqueueSnackbar(progressMessage, {
       variant: 'info',
       persist: true,
-    });
+    })
     try {
-      let result = await promise;
-      closeSnackbar(id);
+      let result = await promise
+      closeSnackbar(id)
       if (successMessage) {
-        enqueueSnackbar(successMessage, { variant: 'success' });
+        enqueueSnackbar(successMessage, { variant: 'success' })
       }
       if (onSuccess) {
-        onSuccess(result);
+        onSuccess(result)
       }
     } catch (e) {
-      console.warn(e);
-      closeSnackbar(id);
+      console.warn(e)
+      closeSnackbar(id)
       // enqueueSnackbar(e.message, { variant: 'error' });
       // if (onError) {
       //   onError(e);
       // }
     }
-  };
+  }
 }
 
-
-
 export function notify({
-  message = "",
+  message = '',
   description = undefined as any,
-  txid = "",
-  type = "info",
-  placement = "bottomLeft",
+  txid = '',
+  type = 'info',
+  placement = 'bottomLeft',
 }) {
   if (txid) {
     //   <Link
@@ -144,16 +148,16 @@ export function notify({
     //     View transaction {txid.slice(0, 8)}...{txid.slice(txid.length - 8)}
     //   </Link>
 
-    description = <></>;
+    description = <></>
   }
-  (notification as any)[type]({
-    message: <span style={{ color: "black" }}>{message}</span>,
+  ;(notification as any)[type]({
+    message: <span style={{ color: 'black' }}>{message}</span>,
     description: (
-      <span style={{ color: "black", opacity: 0.5 }}>{description}</span>
+      <span style={{ color: 'black', opacity: 0.5 }}>{description}</span>
     ),
     placement,
     style: {
-      backgroundColor: "white",
+      backgroundColor: 'white',
     },
-  });
+  })
 }
