@@ -1,43 +1,43 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { PublicKey } from '@solana/web3.js';
-import BN from 'bn.js';
+import { useCallback, useEffect, useMemo, useState } from 'react'
+import { PublicKey } from '@solana/web3.js'
+import BN from 'bn.js'
 
 export function isValidPublicKey(key) {
   if (!key) {
-    return false;
+    return false
   }
   try {
-    new PublicKey(key);
-    return true;
+    new PublicKey(key)
+    return true
   } catch {
-    return false;
+    return false
   }
 }
 
 export async function sleep(ms) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
+  return new Promise((resolve) => setTimeout(resolve, ms))
 }
 
 export const percentFormat = new Intl.NumberFormat(undefined, {
   style: 'percent',
   minimumFractionDigits: 2,
   maximumFractionDigits: 2,
-});
+})
 
 export function floorToDecimal(
   value: number,
-  decimals: number | undefined | null,
+  decimals: number | undefined | null
 ) {
   return decimals
     ? Math.floor(value * 10 ** decimals) / 10 ** decimals
-    : Math.floor(value);
+    : Math.floor(value)
 }
 
 export function roundToDecimal(
   value: number,
-  decimals: number | undefined | null,
+  decimals: number | undefined | null
 ) {
-  return decimals ? Math.round(value * 10 ** decimals) / 10 ** decimals : value;
+  return decimals ? Math.round(value * 10 ** decimals) / 10 ** decimals : value
 }
 
 export function getDecimalCount(value): number {
@@ -46,140 +46,140 @@ export function getDecimalCount(value): number {
     Math.floor(value) !== value &&
     value.toString().includes('.')
   )
-    return value.toString().split('.')[1].length || 0;
+    return value.toString().split('.')[1].length || 0
   if (
     !isNaN(value) &&
     Math.floor(value) !== value &&
     value.toString().includes('e')
   )
-    return parseInt(value.toString().split('e-')[1] || '0');
-  return 0;
+    return parseInt(value.toString().split('e-')[1] || '0')
+  return 0
 }
 
 export function divideBnToNumber(numerator: BN, denominator: BN): number {
-  const quotient = numerator.div(denominator).toNumber();
-  const rem = numerator.umod(denominator);
-  const gcd = rem.gcd(denominator);
-  return quotient + rem.div(gcd).toNumber() / denominator.div(gcd).toNumber();
+  const quotient = numerator.div(denominator).toNumber()
+  const rem = numerator.umod(denominator)
+  const gcd = rem.gcd(denominator)
+  return quotient + rem.div(gcd).toNumber() / denominator.div(gcd).toNumber()
 }
 
 export function getTokenMultiplierFromDecimals(decimals: number): BN {
-  return new BN(10).pow(new BN(decimals));
+  return new BN(10).pow(new BN(decimals))
 }
 
-const localStorageListeners = {};
+const localStorageListeners = {}
 
 export function useLocalStorageStringState(
   key: string,
-  defaultState: string | null = null,
+  defaultState: string | null = null
 ): [string | null, (newState: string | null) => void] {
-  const state = localStorage.getItem(key) || defaultState;
+  const state = localStorage.getItem(key) || defaultState
 
-  const [, notify] = useState(key + '\n' + state);
+  const [, notify] = useState(key + '\n' + state)
 
   useEffect(() => {
     if (!localStorageListeners[key]) {
-      localStorageListeners[key] = [];
+      localStorageListeners[key] = []
     }
-    localStorageListeners[key].push(notify);
+    localStorageListeners[key].push(notify)
     return () => {
       localStorageListeners[key] = localStorageListeners[key].filter(
-        (listener) => listener !== notify,
-      );
+        (listener) => listener !== notify
+      )
       if (localStorageListeners[key].length === 0) {
-        delete localStorageListeners[key];
+        delete localStorageListeners[key]
       }
-    };
-  }, [key]);
+    }
+  }, [key])
 
   const setState = useCallback<(newState: string | null) => void>(
     (newState) => {
-      const changed = state !== newState;
+      const changed = state !== newState
       if (!changed) {
-        return;
+        return
       }
 
       if (newState === null) {
-        localStorage.removeItem(key);
+        localStorage.removeItem(key)
       } else {
-        localStorage.setItem(key, newState);
+        localStorage.setItem(key, newState)
       }
       localStorageListeners[key]?.forEach((listener) =>
-        listener(key + '\n' + newState),
-      );
+        listener(key + '\n' + newState)
+      )
     },
-    [state, key],
-  );
+    [state, key]
+  )
 
-  return [state, setState];
+  return [state, setState]
 }
 
 export function useLocalStorageState<T = any>(
   key: string,
-  defaultState: T | null = null,
+  defaultState: T | null = null
 ): [T, (newState: T) => void] {
   let [stringState, setStringState] = useLocalStorageStringState(
     key,
-    JSON.stringify(defaultState),
-  );
+    JSON.stringify(defaultState)
+  )
   return [
     useMemo(() => stringState && JSON.parse(stringState), [stringState]),
     (newState) => setStringState(JSON.stringify(newState)),
-  ];
+  ]
 }
 
 export function useEffectAfterTimeout(effect, timeout) {
   useEffect(() => {
-    const handle = setTimeout(effect, timeout);
-    return () => clearTimeout(handle);
-  });
+    const handle = setTimeout(effect, timeout)
+    return () => clearTimeout(handle)
+  })
 }
 
 export function useListener(emitter, eventName) {
-  const [, forceUpdate] = useState(0);
+  const [, forceUpdate] = useState(0)
   useEffect(() => {
-    const listener = () => forceUpdate((i) => i + 1);
-    emitter.on(eventName, listener);
-    return () => emitter.removeListener(eventName, listener);
-  }, [emitter, eventName]);
+    const listener = () => forceUpdate((i) => i + 1)
+    emitter.on(eventName, listener)
+    return () => emitter.removeListener(eventName, listener)
+  }, [emitter, eventName])
 }
 
 export function abbreviateAddress(address: PublicKey, size = 4) {
-  const base58 = address.toBase58();
-  return base58.slice(0, size) + '…' + base58.slice(-size);
+  const base58 = address.toBase58()
+  return base58.slice(0, size) + '…' + base58.slice(-size)
 }
 
 export function isEqual(obj1, obj2, keys) {
   if (!keys && Object.keys(obj1).length !== Object.keys(obj2).length) {
-    return false;
+    return false
   }
-  keys = keys || Object.keys(obj1);
+  keys = keys || Object.keys(obj1)
   for (const k of keys) {
     if (obj1[k] !== obj2[k]) {
       // shallow comparison
-      return false;
+      return false
     }
   }
-  return true;
+  return true
 }
 
 export function flatten(obj, { prefix = '', restrictTo }) {
-  let restrict = restrictTo;
+  let restrict = restrictTo
   if (restrict) {
-    restrict = restrict.filter((k) => obj.hasOwnProperty(k));
+    restrict = restrict.filter((k) => obj.hasOwnProperty(k))
   }
-  const result = {};
-  (function recurse(obj, current, keys) {
-    (keys || Object.keys(obj)).forEach((key) => {
-      const value = obj[key];
-      const newKey = current ? current + '.' + key : key; // joined key with dot
+  const result = {}
+  ;(function recurse(obj, current, keys) {
+    ;(keys || Object.keys(obj)).forEach((key) => {
+      const value = obj[key]
+      const newKey = current ? current + '.' + key : key // joined key with dot
       if (value && typeof value === 'object') {
         // @ts-ignore
-        recurse(value, newKey); // nested object
+        recurse(value, newKey) // nested object
       } else {
-        result[newKey] = value;
+        result[newKey] = value
       }
-    });
-  })(obj, prefix, restrict);
-  return result;
+    })
+  })(obj, prefix, restrict)
+  return result
 }
